Added sine and cosine to evaluation.

+++ library.c
@@ -47,7 +47,7 @@ int rgb(int r, int g, int b){
 double distance(int x1, int y1, int x2, int y2){
   return sqrt(pow(x2 - x1, 2) + pow(y2 - y1, 2));
 }
-double degrees_to_radians(int degrees){
+double degrees_to_radians(double degrees){
   return (M_PI * degrees) / 180.0;
 }
 int radians_to_degrees(double radians){
